requirements.txt
Flask
flask-cors
psycopg2-binary
scikit-learn
pandas
numpy
pip install -r requirements.txt